+++ FE_Onboarding/src/Modules/Admin/components/ContractServicesDashboard.module.css
@@ -1,0 +1,50 @@
+.container {
+  padding: 30px;
+  background-color: #f7f9fc;
+  min-height: 100vh;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 25px;
+  padding-bottom: 15px;
+  border-bottom: 2px solid #e0e0e0;
+}
+
+.title {
+  font-size: 1.8rem;
+  color: var(--bdo-charcoal, #333);
+  display: flex;
+  align-items: center;
+  font-weight: 600;
+}
+
+.icon {
+  margin-right: 10px;
+  color: var(--bdo-red, #cc0000);
+}
+
+.description {
+  color: #666;
+  margin-bottom: 20px;
+  font-size: 1rem;
+}
+
+.loadingMessage,
+.savingMessage {
+  padding: 15px;
+  text-align: center;
+  font-size: 1.1rem;
+  font-weight: 500;
+  color: var(--bdo-red, #cc0000);
+  background-color: #fff;
+  border-radius: 8px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
+}
+
+.savingMessage {
+  margin-top: 15px;
+  color: #4caf50; /* Un color de éxito/guardado */
+}
